refactor: [frontend] [APP: todo] GChartHelper 클래스 생성 및 리팩토링
