+++ frontend/app/src/components/TodoApp.vue
@@ -37,8 +37,9 @@
 <script>
 import AddTodo from './AddTodo.vue';
 import TodoList from './TodoList.vue';
-import TodoChart from "./TodoChart.vue";
 import StopWatch from './StopWatch.vue';
+import TodoChart from './TodoChart.vue';
+import GChartHelper from '../classes/GChartHelper.js';
 
 export default {
   components: {
@@ -64,9 +65,10 @@ export default {
       return todos.filter(t => t.completed !== status);
     },
     setTodoList() {
-        this.completed_todos = this.getTodosByCompleted(this.todos, "N");
-        this.uncompleted_todos = this.getTodosByCompleted(this.todos, "Y");
-        this.chart_data = this.getChartData(this.todos);
+      this.completed_todos = this.getTodosByCompleted(this.todos, "N");
+      this.uncompleted_todos = this.getTodosByCompleted(this.todos, "Y");
+      const gchart = new GChartHelper(this.$moment, this.todos);
+      this.chart_data = gchart.getChartData(this.todos);
     },
     async getTodos() {
       try {
@@ -96,55 +98,6 @@ export default {
     async changeTodo(todo) {
       this.setTodoList();
     },
-
-
-    getChartData(todos) {
-        const columns = [
-            { type: 'string', id: 'todo' },
-            { type: 'string', id: 'content' },
-            { type: 'date', id: 'Start' },
-            { type: 'date', id: 'End' },
-        ];
-
-        const rows = []
-        todos.forEach((todo) => {
-            const started_at = this.cvtTimeStringToHMSDate(todo.started_at);
-            const ended_at = this.cvtTimeStringToHMSDate(todo.ended_at);
-            if (this.isValidTodoForChart(todo)) {
-                rows.push([
-                    this.getTimeTitleByDate(todo.started_at), todo.content, started_at, ended_at
-                ]);
-            }
-        });
-        return [columns, ...rows];
-    },
-    isValidTodoForChart(todo) {
-        if (!todo.started_at) {
-            return false;
-        }
-        if (!todo.ended_at) {
-            return false;
-        }
-        if (todo.completed === 'N') {
-            return false;
-        }
-        return true;
-    },
-    getTimeTitleByDate(date) {
-        const time_parts = date.split('T')
-        const ymd = time_parts[0].split('-')
-        return `${ymd[1]}-${ymd[2]}`
-    },
-    cvtTimeStringToHMSDate(timeString) {
-      if (!timeString) {
-          return null
-      }
-      const date = this.$moment(timeString).toDate();
-      const h = date.getHours();
-      const m = date.getMinutes();
-      const s = date.getSeconds();
-      return new Date(0, 0, 0, h, m, s);
-    }
   }
 };
 </script>
